feat: create new dashboard shell with persistent layout, preserve legacy at /legacy
- Add new Dashboard.tsx with persistent shell layout architecture:
  - Full-width top bar with logo and content area
  - Collapsible sidebar (220px expanded, 64px collapsed)
  - Main content area with placeholder
- Create Dashboard.css with styling using design tokens
- Rename original Dashboard to DashboardLegacy and preserve at /legacy route
- Update design tokens in tokens.css:
  - New radius values (6px, 8px, 12px)
  - Universal gap tokens (--gap-container, --gap-element, --gap-tight)
  - Full color palette for theme consistency
- Update App.tsx routing to handle / and /legacy routes
