+++ src/App.tsx
@@ -1,5 +1,6 @@
 import { useState, useCallback, useEffect } from 'react';
 import { Dashboard } from './components/Dashboard';
+import { DashboardLegacy } from './components/DashboardLegacy';
 import { ErrorBoundary } from './components/ErrorBoundary';
 import { AuthProvider, useAuthContext } from './contexts/AuthContext';
 import { ToastProvider, useToastContext } from './contexts/ToastContext';
@@ -22,7 +23,7 @@ import './styles/help.css';
 // ROUTE TYPES
 // ============================================================================
 
-type AppRoute = 'dashboard' | 'account' | 'help';
+type AppRoute = 'dashboard' | 'legacy' | 'account' | 'help';
 
 function AppContent() {
   const [parsedData, setParsedData] = useState<ParsedDRFFile | null>(null);
@@ -31,8 +32,14 @@ function AppContent() {
   const [showWarnings, setShowWarnings] = useState(true);
   const [modalOpen] = useState(false);
 
-  // Routing state
-  const [currentRoute, setCurrentRoute] = useState<AppRoute>('dashboard');
+  // Routing state - check URL on initial load
+  const getInitialRoute = (): AppRoute => {
+    if (window.location.pathname === '/legacy') {
+      return 'legacy';
+    }
+    return 'dashboard';
+  };
+  const [currentRoute, setCurrentRoute] = useState<AppRoute>(getInitialRoute);
 
   // Auth state
   const { isAuthenticated, isLoading: authLoading } = useAuthContext();
@@ -257,30 +264,39 @@ function AppContent() {
     );
   }
 
-  // Show main dashboard
+  // Show legacy dashboard at /legacy route
+  if (currentRoute === 'legacy') {
+    return (
+      <ErrorBoundary onReset={handleFullReset}>
+        {/* Disclaimer Banner - shows on first visit */}
+        <DisclaimerBanner onViewFull={handleViewFullDisclaimer} />
+
+        <DashboardLegacy
+          parsedData={parsedData}
+          isLoading={isLoading}
+          validationWarnings={validationWarnings}
+          showWarnings={showWarnings}
+          onParsed={handleParsed}
+          onDismissWarnings={handleDismissWarnings}
+          raceState={raceState}
+          onOpenLegalModal={(type: LegalContentType) => {
+            setLegalModalType(type);
+            setLegalModalOpen(true);
+          }}
+          onNavigateToAccount={navigateToAccount}
+          onNavigateToHelp={navigateToHelp}
+        />
+
+        {/* Legal Modal - shared across app */}
+        <LegalModal type={legalModalType} isOpen={legalModalOpen} onClose={handleCloseLegalModal} />
+      </ErrorBoundary>
+    );
+  }
+
+  // Show new dashboard shell at root route
   return (
     <ErrorBoundary onReset={handleFullReset}>
-      {/* Disclaimer Banner - shows on first visit */}
-      <DisclaimerBanner onViewFull={handleViewFullDisclaimer} />
-
-      <Dashboard
-        parsedData={parsedData}
-        isLoading={isLoading}
-        validationWarnings={validationWarnings}
-        showWarnings={showWarnings}
-        onParsed={handleParsed}
-        onDismissWarnings={handleDismissWarnings}
-        raceState={raceState}
-        onOpenLegalModal={(type: LegalContentType) => {
-          setLegalModalType(type);
-          setLegalModalOpen(true);
-        }}
-        onNavigateToAccount={navigateToAccount}
-        onNavigateToHelp={navigateToHelp}
-      />
-
-      {/* Legal Modal - shared across app */}
-      <LegalModal type={legalModalType} isOpen={legalModalOpen} onClose={handleCloseLegalModal} />
+      <Dashboard />
     </ErrorBoundary>
   );
 }
