+++ src/styles/tokens.css
@@ -3,16 +3,12 @@
    ============================================ */
 
 :root {
-  /* ============================================
-     Border Radius Tokens
-     ============================================ */
-  --radius-sm: 4px;   /* badges, chips */
-  --radius-md: 6px;   /* buttons, inputs */
-  --radius-lg: 8px;   /* cards, panels, modals */
+  /* Border radius - balanced, professional */
+  --radius-sm: 6px;
+  --radius-md: 8px;
+  --radius-lg: 12px;
 
-  /* ============================================
-     Spacing Tokens (8px grid)
-     ============================================ */
+  /* Spacing - 8px base grid */
   --space-1: 4px;
   --space-2: 8px;
   --space-3: 12px;
@@ -22,11 +18,27 @@
   --space-7: 48px;
   --space-8: 64px;
 
-  /* ============================================
-     Gap Tokens
-     ============================================ */
-  --gap-panel: 16px;    /* between major panels */
-  --gap-card: 16px;     /* between cards */
-  --gap-element: 8px;   /* between elements within a card */
-  --gap-tight: 4px;     /* compact spacing */
+  /* Universal gaps - THESE ARE THE ONLY VALUES FOR CONTAINER SPACING */
+  --gap-container: 16px; /* Between and inside all cards, panels, sections */
+  --gap-element: 8px; /* Between elements within a card (label to input, icon to text) */
+  --gap-tight: 4px; /* Compact spacing for dense UI elements */
+
+  /* Colors - matching existing theme */
+  --color-base: #0a0a0b;
+  --color-card: #0f0f10;
+  --color-elevated: #1a1a1c;
+  --color-border: #2a2a2c;
+  --color-border-prominent: #3a3a3c;
+
+  --color-primary: #19abb5;
+  --color-primary-hover: #1992a1;
+  --color-primary-muted: rgba(25, 171, 181, 0.15);
+
+  --color-text-primary: #eeeff1;
+  --color-text-secondary: #b4b4b6;
+  --color-text-tertiary: #6e6e70;
+
+  --color-success: #10b981;
+  --color-warning: #f59e0b;
+  --color-error: #ef4444;
 }
